move the update task into its own file
